MAINT: stats: move _contains_nan function to _lib._util.py
