+++ scipy/stats/_axis_nan_policy.py
@@ -9,6 +9,7 @@ import scipy.stats
 import scipy.stats._stats_py
 from functools import wraps
 from scipy._lib._docscrape import FunctionDoc, Parameter
+from scipy._lib._util import _contains_nan
 import inspect
 
 
@@ -475,8 +476,7 @@ def _axis_nan_policy_factory(tuple_to_result, default_axis=0,
                 # Addresses nan_policy == "raise"
                 contains_nans = []
                 for sample in samples:
-                    contains_nan, _ = (
-                        scipy.stats._stats_py._contains_nan(sample, nan_policy))
+                    contains_nan, _ = _contains_nan(sample, nan_policy)
                     contains_nans.append(contains_nan)
 
                 # Addresses nan_policy == "propagate"
@@ -523,8 +523,7 @@ def _axis_nan_policy_factory(tuple_to_result, default_axis=0,
             x = _broadcast_concatenate(samples, axis)
 
             # Addresses nan_policy == "raise"
-            contains_nan, _ = (
-                scipy.stats._stats_py._contains_nan(x, nan_policy))
+            contains_nan, _ = _contains_nan(x, nan_policy)
 
             if vectorized and not contains_nan and not sentinel:
                 res = hypotest_fun_out(*samples, axis=axis, **kwds)
